creation of some overrinding functions on actetechnique dao file also fixing expertise logic jpa

+++ src/main/java/com/example/platform_tele_expertise_medicale/controller/SpecialistProfileServlet.java
@@ -69,7 +69,7 @@ public class SpecialistProfileServlet extends HttpServlet {
         try {
             Utilisateur user = (Utilisateur) session.getAttribute("user");
             
-            Long specialiteId = Long.parseLong(request.getParameter("specialiteId"));
+            Integer specialiteId = Integer.parseInt(request.getParameter("specialiteId"));
             BigDecimal tarif = new BigDecimal(request.getParameter("tarif"));
             
             Specialite specialite = specialiteDAO.findById(specialiteId);
